requirements.txt
torch
transformers
pillow
scikit-learn
streamlit